refactor out UI components
